+++ tests/testthat/test_postgis_RPostgres.R
@@ -67,7 +67,7 @@ test_that("sf can write units to database (#264)", {
     skip_if_not(can_con(pg), "could not connect to postgis database")
     ptsu <- pts
     ptsu[["u"]] <- ptsu[["cadmium"]]
-    units(ptsu[["u"]]) <- units::make_unit("km")
+    units(ptsu[["u"]]) <- units::as_units("km")
     expect_silent(st_write(ptsu, pg, "sf_units__", overwrite = TRUE))
     r <- st_read(pg, "sf_units__")
     expect_is(r[["u"]], "numeric")
@@ -89,6 +89,37 @@ test_that("validates arguments", {
     expect_error(st_read(pg, "sf_meuse__", random_arg = "a"), "Unused arguments:")
     expect_error(st_read(pg, "sf_meuse__", table = "a"), "`layer` rather than `table`")
     expect_error(st_read(pg, "sf_meuse__", table = "a", x = 1, y = 2), "`layer` rather than `table`")
+})
+
+test_that("sf can write non-sf tables with geometries", {
+    skip_if_not(can_con(pg), "could not connect to postgis database")
+    df <- as.data.frame(pts)
+    expect_silent(st_write(df, pg, "df"))
+    expect_silent(dfx <- st_read(pg, "df"))
+    expect_equal(df[["geometry"]], dfx[["geometry"]])
+    expect_silent(DBI::dbRemoveTable(pg, "df"))
+})
+
+test_that("sf can write non-sf tables with multiple geometries", {
+    skip_if_not(can_con(pg), "could not connect to postgis database")
+    df <- as.data.frame(pts)
+    df$geography <- st_transform(df$geometry, 4326)
+    expect_silent(st_write(df, pg, "df"))
+    expect_silent(dfx <- st_read(pg, "df"))
+    expect_equal(df[["geometry"]], dfx[["geometry"]])
+    expect_equal(df[["geography"]], dfx[["geography"]])
+    expect_silent(DBI::dbRemoveTable(pg, "df"))
+})
+
+test_that("tidy workflow can write multiple geometries", {
+    skip_if_not(can_con(pg), "could not connect to postgis database")
+    df <- tibble::as_tibble(pts)
+    df <- dplyr::mutate(df, geography = st_transform(geometry, 4326))
+    expect_silent(write_sf(df, pg, "df"))
+    on.exit(DBI::dbRemoveTable(pg, "df"))
+    expect_silent(dfx <- read_sf(pg, "df"))
+    expect_equal(df[["geometry"]], dfx[["geometry"]])
+    expect_equal(df[["geography"]], dfx[["geography"]])
 })
 
 test_that("sf can preserve types (#592)", {
@@ -115,12 +146,10 @@ test_that("sf can preserve types (#592)", {
 })
 
 test_that("can write to other schema", {
-	# skip_if_not(FALSE) # tmp switch off -- EJP
     skip_if_not(can_con(pg), "could not connect to postgis database")
     try(DBI::dbSendQuery(pg, "CREATE SCHEMA sf_test__;"), silent = TRUE)
     q <- "SELECT schema_name FROM information_schema.schemata WHERE schema_name = 'sf_test__';"
     suppressWarnings(could_schema <- DBI::dbGetQuery(pg, q) %>% nrow() > 0)
-
     skip_if_not(could_schema, "Could not create schema (might need to run 'GRANT CREATE ON DATABASE postgis TO <user>')")
     expect_error(st_write(pts, pg, Id(schema = "public", table = "sf_meuse__")), "exists")
     expect_silent(st_write(pts, pg, Id(schema = "sf_test__", table = "sf_meuse__")))
@@ -304,8 +333,15 @@ test_that("Can safely manipulate crs", {
     # udpate
     expect_message(set_postgis_crs(pg, new_srid), "Inserted local crs")
     new_srid$proj4string <- crs2$proj4string
-    expect_error(set_postgis_crs(pg, new_srid), "already exists")
-    expect_message(set_postgis_crs(pg, new_srid, update = TRUE), "Inserted local crs")
+    expect_warning(
+        expect_error(set_postgis_crs(pg, new_srid), "already exists"),
+        "GDAL Error 6: EPSG"
+    )
+    expect_warning(
+        expect_message(set_postgis_crs(pg, new_srid, update = TRUE), "Inserted local crs"),
+        "GDAL Error 6: EPSG"
+    )
+
 })