+++ tests/testthat/test_read.R
@@ -41,7 +41,7 @@ test_that("guess_driver works on extensions", {
   # for repeatability, this is how I turned the list to tests
   # `^"(\w+)" = ("[\w\s]+"),?`
   # to `expect_equal(guess_driver("nc.\1"), c("\1" = \2))`
-  
+
   expect_equal(guess_driver("nc.bna"), c("bna" = "BNA"))
   expect_equal(guess_driver("nc.csv"), c("csv" = "CSV"))
   expect_equal(guess_driver("nc.e00"), c("e00" = "AVCE00"))
@@ -51,7 +51,7 @@ test_that("guess_driver works on extensions", {
   expect_equal(guess_driver("nc.gmt"), c("gmt" = "GMT"))
   expect_equal(guess_driver("nc.gpkg"), c("gpkg" = "GPKG"))
   expect_equal(guess_driver("nc.gps"), c("gps" = "GPSBabel"))
-  expect_equal(guess_driver("nc.gtm"), c("gtm" = "GPSTrackMaker"))   
+  expect_equal(guess_driver("nc.gtm"), c("gtm" = "GPSTrackMaker"))
   expect_equal(guess_driver("nc.gxt"), c("gxt" = "Geoconcept"))
   expect_equal(guess_driver("nc.kml"), c("kml" = "KML"))
   expect_equal(guess_driver("nc.jml"), c("jml" = "JML"))
@@ -66,7 +66,7 @@ test_that("guess_driver works on extensions", {
   expect_equal(guess_driver("nc.vdv"), c("vdv" = "VDV"))
   expect_equal(guess_driver("nc.xls"), c("xls" = "xls"))
   expect_equal(guess_driver("nc.xlsx"), c("xlsx" = "XLSX"))
-  
+
   # unsuported
   expect_equal(guess_driver("nc.notsupported"), NA)
 })
@@ -75,7 +75,7 @@ test_that("guess_driver works on suffixes", {
   # for repeatability, this is how I turned the list to tests
   # `^"(\w+)" = ("[\w\s]+"),?`
   # to `expect_equal(guess_driver("nc.\1"), c("\1" = \2))`
-  
+
   expect_equal(guess_driver("couchdb:nc"), c("couchdb" = "CouchDB"))
   expect_equal(guess_driver("db2odbc:nc"), c("db2odbc" = "DB2ODBC"))
   expect_equal(guess_driver("dods:nc"), c("dods" = "DODS"))
@@ -86,7 +86,7 @@ test_that("guess_driver works on suffixes", {
   expect_equal(guess_driver("odbc:nc"), c("odbc" = "ODBC"))
   expect_equal(guess_driver("pg:nc"), c("pg" = "PostgreSQL"))
   expect_equal(guess_driver("sde:nc"), c("sde" = "SDE"))
-  
+
   # upper case
   expect_equal(guess_driver("CouchDB:nc"), c("couchdb" = "CouchDB"))
   expect_equal(guess_driver("db2ODBC:nc"), c("db2odbc" = "DB2ODBC"))
@@ -98,7 +98,7 @@ test_that("guess_driver works on suffixes", {
   expect_equal(guess_driver("ODBC:nc"), c("odbc" = "ODBC"))
   expect_equal(guess_driver("PG:nc"), c("pg" = "PostgreSQL"))
   expect_equal(guess_driver("SDE:nc"), c("sde" = "SDE"))
-  
+
   # unsuported
   expect_equal(guess_driver("notsupported:nc"), NA)
 })
@@ -108,7 +108,7 @@ test_that("weird names are supported", {
   expect_equal(guess_driver("pg:nc.shp.e00"), c("pg" = "PostgreSQL"))
   expect_equal(guess_driver("nc.shp.e00"), c("e00" = "AVCE00"))
   expect_equal(guess_driver("couchdb:shp"), c("couchdb" = "CouchDB"))
-  
+
   expect_equal(guess_driver("notsupported:nc.shp"), NA)
   expect_equal(guess_driver("notsupported"), NA)
 })
@@ -118,7 +118,7 @@ test_that("driver utils work", {
   expect_true(is_driver_available("shp", data.frame(name = c("shp"))))
   expect_false(is_driver_available("shp", data.frame(name = c("x", "y", "z"))))
   expect_false(is_driver_available("shp", data.frame(name = c("x", "y", "z"))))
-  
+
   expect_error(is_driver_can("shp", data.frame(name = c("x", "y", "shp")), operation = "nothing"))
   expect_error(is_driver_can("shp", data.frame(name = c("x", "y")), operation = "nothing"))
   expect_true(is_driver_can("shp", data.frame(name = c("x", "y", "shp"), write = rep(TRUE, 3)), operation = "write"))
@@ -130,23 +130,23 @@ test_that("guess_driver_can_write", {
   expect_error(guess_driver_can_write("x.not", c("nothing" = "nothing")), "not available")
   expect_equal(guess_driver_can_write("x.csv"), c("csv" = "CSV"))
   expect_equal(guess_driver_can_write("c:/x.csv"), c("csv" = "CSV"))
-  
+
   expect_error(guess_driver_can_write("x.unsuported"), "Could not guess driver")
   expect_error(guess_driver_can_write("unsuported:x"), "Could not guess driver")
 })
 
-test_that("driver operations", {  
+test_that("driver operations", {
   # These tests are driver specifics to GDAL version and OS.
   expect_error(guess_driver_can_write("x.e00"), "cannot write")
   expect_error(guess_driver_can_write("x.gdb"), "cannot write")
-  
+
   expect_equal(guess_driver_can_write("x.geojson"), c("geojson" = "GeoJSON"))
   expect_equal(guess_driver_can_write("x.csv"), c("csv" = "CSV"))
   expect_equal(guess_driver_can_write("x.gml"), c("gml" = "GML"))
 })
 
 test_that("guess driver on windows with backslashes (#127)", {
-    expect_identical(guess_driver("c:\\Temp\\this.shp"), 
+    expect_identical(guess_driver("c:\\Temp\\this.shp"),
                      guess_driver("c:/Temp/this.shp"))
 })
 
@@ -169,12 +169,18 @@ test_that("we get a warning when not specifying one of multiple layers", {
 })
 
 test_that("reading non-spatial table works", {
-	skip_if_not(sf_extSoftVersion()[["GDAL"]] >= "2.2.0") # error on OSX for 2.1.3
-	
-	expect_warning(st_read(system.file("gpkg/nospatial.gpkg", package = "sf")),
-				   "no simple feature geometries present")
-	expect_is(st_read(system.file("gpkg/nospatial.gpkg", package = "sf")),
-			  "data.frame")
-	expect_is(read_sf(system.file("gpkg/nospatial.gpkg", package = "sf")),
-			  "tbl_df")
+    skip_if_not(sf_extSoftVersion()[["GDAL"]] >= "2.2.0") # error on OSX for 2.1.3
+
+    expect_warning(st_read(system.file("gpkg/nospatial.gpkg", package = "sf")),
+                   "no simple feature geometries present")
+    expect_warning(
+        expect_is(st_read(system.file("gpkg/nospatial.gpkg", package = "sf")),
+                  "data.frame"),
+        "no simple feature geometries"
+    )
+    expect_warning(
+        expect_is(read_sf(system.file("gpkg/nospatial.gpkg", package = "sf")),
+              "tbl_df"),
+        "no simple feature geometries"
+    )
 })